Add go mod support and fix compilation on windows for golang 1.11
