+++ .gitignore
@@ -1,1 +1,2 @@
 apksig_for_tests
+vendor/
