+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/avast/apkverifier
+
+go 1.10
+
+require github.com/avast/apkparser v0.0.0-20190516101250-3b8c5efcb6a9
